Fix bluetooth.find_service under Windows
The commit reverts 0091913, which introduced multiple bugs in the
codebase.
Indeed, adding the line #define UNICODE to a Windows C source has the
effect of switching all WIndows API functions from the (deprecated) ASCII
version to the new UTF-16 version. Porting a source to UTF-16 cannot be
done just with a find & replace, because it shall be taken into account that
anything going to manipulate a string will now work with wide
characters.

Moreover, the port to Python 3 of the source _msbt.c does not take
properly into account the fact that in Python3 str and bytes are two
different types: str are UTF-8 by default now, while bytes is the right
type to handle generic byte arrays.

To sum up:
 - parsing a C byte array to a Python3 string could fail, due to errors
 during UTF-8 decoding. C byte arrays shall be parsed to the new bytes type.
  - parsing a UTF-16 wchar_t as a Python3 string and the other way around
  will fail, if no re-encoding is done

The same bugs likely affect the entire _msbt.c source and shall be
fixed ASAP in order to actually support WIndows.